Added structure for hash operations decalarations: Fix Rpush and test expire properly!

+++ src/RedisDatabase.cpp
@@ -204,7 +204,7 @@ bool RedisDatabase::lindex(const std::string& key, int index, std::string& value
   const auto& lst = it->second;
   if (index < 0)
     index = lst.size() + index; //convert negative index
-  if (index < 0 || static_cast<ssize_t>(index) >= lst.size()) 
+  if (index < 0 || index >= static_cast<int>(lst.size())) 
     return false;
   value = lst[index];
   return true;
@@ -219,7 +219,7 @@ bool RedisDatabase::lset(const std::string& key, int index, const std::string& v
   auto& lst = it->second;
   if (index < 0)  
     index = lst.size() + index; //convert negative index
-  if (index < 0 || static_cast<ssize_t>(index) >= lst.size()) 
+  if (index < 0 || index >= static_cast<int>(lst.size())) 
     return false;   
   
   lst[index] = value;
